Move search view into own file and style changes
