+++ demos/voice-search/src/index.css
@@ -24,7 +24,6 @@ body {
   font-family: 'Inter', sans-serif;
   font-weight: 400;
   line-height: 1.25;
-  overflow: hidden;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
